LayoutTests/http/tests/inspector/network/network-sidebar-width-expected.txt

CONSOLE MESSAGE: line 13: Done.
Tests network panel sidebar width with and without preference set.

Bug 65256
Sidebar width: 200
Sidebar width: 242
